Commited missing fix to grpThread grouping time getter.

+++ Clustering/groupingthread.h
@@ -49,7 +49,7 @@ public:
 
     std::vector<std::shared_ptr<cluster>> clusters;
 
-    long getGroupingTime();
+    double getGroupingTime();
 
 signals:
